chore: change project structure to start at src and configure absolute imports using @ prefix
