+++ src/components/FirstComponent.tsx
@@ -1,0 +1,9 @@
+const FirstComponent = () => {
+  return (
+    <div>
+      <h1>Test</h1>
+    </div>
+  )
+}
+
+export default FirstComponent
