+++ src/pages/index.tsx
@@ -1,0 +1,22 @@
+import FirstComponent from '@components/FirstComponent'
+import Head from 'next/head'
+import Image from 'next/image'
+import styles from '../../styles/Home.module.css'
+
+export default function Home() {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <main className={styles.main}>
+        <FirstComponent />
+      </main>
+
+      <footer className={styles.footer}>Test</footer>
+    </div>
+  )
+}
